% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourPolys-package.R
\docType{package}
\name{contourPolys-package}
\alias{contourPolys}
\alias{contourPolys-package}
\title{contourPolys: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item R Core Team (copyright plot3d.c code in graphics package, reproduced here)
}

}
\keyword{internal}
