% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fcontour}
\alias{fcontour}
\title{Filled contour}
\usage{
fcontour(x, y, z, c)
}
\arguments{
\item{x}{vector x coordinate}

\item{y}{vector x coordinate}

\item{z}{vector z matrix}
}
\description{
Filled contour
}
