% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fcontour_sf}
\alias{fcontour_sf}
\title{Filled contour}
\usage{
fcontour_sf(x, y, z, c)
}
\arguments{
\item{x}{vector x coordinate}

\item{y}{vector x coordinate}

\item{z}{vector z matrix}
}
\description{
Filled contour
}
\examples{
z <- as.matrix(volcano)
 y <- seq_len(ncol(z))
 x <- seq_len(nrow(z))
 levels <- pretty(range(z), n =7)
 fc <- fcontour_sf(x, y, z, c = levels)
 g <- purrr::map(fc[[1]], ~sf::st_polygon(list(.x)))
 ik <- unlist(fc[[2]])
 ## here we need to trim, ignore any 0-area polygons in C++
 bad <- !unlist(lapply(g, st_is_valid))
 library(dplyr)
 x <- st_sf(geometry = st_sfc(g[!bad]), kk = ik[!bad]) \%>\% group_by(kk) \%>\% summarize() \%>\% st_cast("MULTIPOLYGON")
}
