% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sfill_contour}
\alias{sfill_contour}
\alias{sfill_contour.BasicRaster}
\title{Title}
\usage{
sfill_contour(x, levels = NULL, ..., maxpixels = 2e+16)

\method{sfill_contour}{BasicRaster}(x, levels = NULL, ...,
  maxpixels = 2^15)
}
\arguments{
\item{x}{}

\item{levels}{}

\item{...}{}

\item{maxpixels}{}
}
\description{
Title
}
